@@ -12,15 +10,14 @@
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
-        <version>2.3.4.RELEASE</version>
+        <version>2.5.6</version>
         <relativePath />
-        <!-- lookup parent from repository -->
     </parent>
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <start-class>com.game.qijeu.QijeuServicesApplication</start-class>
-        <maven.compiler.release>15</maven.compiler.release>
-        <liquibase.version>4.2.0</liquibase.version>
+        <maven.compiler.release>17</maven.compiler.release>
+        <liquibase.version>4.5.0</liquibase.version>
     </properties>
     <dependencies>
         <dependency>
